Re-organised the build of the external projects.
Switched all of them to be built using FetchContent instead of
using git submodules.

Added an external for VecMem directly, so the project would not
need to rely on algebra-plugins to get access to that external.

Wrote a custom configuration for installing the dfe headers together
with the project. Such that it would become possible to export the
targets of this project eventually as well.

Updated all components to use the externals through their "exported names".