Code Re-Organisation, main branch (2021.11.02.) (#145)
* Started with the general cleanup of the build configuration.

Introduced some common CMake code, which could then be used to declare
the libraries and executables of the project in a consistent way.
While also making it possible to export the configuration of the project
correctly.

* Updated the setup of the detray::core library.

Moved all of the headers under the "detray/" directory, and made
sure that all clients would follow this change in naming.

* Fixed the usage of the plugin CACHE variables.

In the new setup the "array plugin" is always active, and
DETRAY_VC_ARRAY_PLUGIN was renamed to simply DETRAY_VC_PLUGIN.

* Updated the setup of the detray::io library.

Moved all of the headers under the "detray/" directory, and made
sure that all clients would follow this change in naming.

In the new setup detray::io and detray::io_csv are two separate
libraries. Though the logic around these will probably need to be
cleaned up a bit.

* Updated the setup of the plugin libraries.

Moved all of the headers under the "detray/" directory, and made
sure that all clients would follow this change in naming.

* Updated the setup of the detray::display library.

Moved all of the headers under the "detray/" directory, and made
sure that all clients would follow this change in naming.

At the same time made some fixes to the build of Matplot++, and to
code in this project that would use it.

* Moved the final elements not meant for the main CMakeLists.txt file, out of there.

* Ran code formatting after the changes in include statements.

* Updates to make the rebased code functional.

Plus a fix to the typo that Joana found. ;-)