Many updates in the project's CMake configuration.
Updated all externals to make it configurable where they would get the
source to build from.

Updated VecMem to version v0.6.0, and Algebra Plugins to v0.3.0.

Updated the setup of Thrust. The build now takes the sources from
git directly, as the packages provided by Thrust are incomplete.
Introduced the setup of the detray::Thrust library, as the library
to link against when using Thrust.

Made numerous fixes in the exported configuration of the project.

Made the installation of library headers optional, to accommodate
the weird setup of detray::io.

Simplified the array and vc_array plugins, as the Algebra Plugins
code allowed for the simplification with its new version.