Adapted the code to the removal of algebra::scalar.
The algebra-plugins code no longer makes a decision during
compilation about the scalar type that it would let its clients
use. The clients need to make that decision themselves.

The Detray code is still making this decision during the CMake
configuration, with the revival of the DETRAY_CUSTOM_SCALARTYPE
cache variable.