Re-wrote how tests and benchmarks would be built in the project.
Most importantly removed all of the individual main() functions from
the test files, to instead link the tests against GTest::gtest_main.
Thereby allowing tests to be grouped into larger executables.

Moved the CMake functions that set up the tests and benchmarks to
cmake/detray-functions.cmake, like it is done in the other projects
as well.