feat: typed indices (#207)
Introduce a dedicated class to deal with indexing that needs to resolve type. This way the mask_store does not need to use a pair or tuple for ranged links. Also, since the surface finder infrastructure will mimic that of the masks, it will enforce more prudence in building those links