exp: trust based navigation (#220)
In this PR, the navigator and the steppers are adapted to use a fully trust_level based communication. The navigator's intersection cache is updated according to different trust levels, which are set by a stepper.
The stepper now sets the distance to the next target at the beginning of every step, before the step size adjustment of the RK, discarding the previous step size. In cases where the magnetic field is not strictly aligned in z direction, the step size is additionally capped by a max step size to force a more fine-grained re-evaluation of the navigation state.
The host side propagation is also now tested against a number of B-field configurations. Some trivial changes were made in the inspector while debugging and the navigator's unittest had to be adapted all over again. The execution speed of the propagation of the different algebra-plugins is now the same on my laptop, while also for me Eigen was previously the fastest.