fix: b field access #238
Using a reference to the B-field caused undefined behavior in the host-side rk-stepper test when using the array plugin. Using pointers works, but I don't really know what the lifetimes of the stepper and the B-field are supposed to be, so lets keep it as it was and copy the B-field, until we know what the B-field access will look like eventually.