feat: add pathlimit aborter (#236)
Adds a pathlimit_aborter as an actor in the actor_chain and removes the path_limit checking from the steppers. Also updates the CUDA stepper and propagator unittests to use constrained stepping. The detray and CUDA propagator tests are furthermore extended to use a pathlimit_aborter, although with a limit that should not prevent the propagation of the entire toy detector.