feat: add helix surface intersectiors for particle gun (#264)
Merges the ray gun and helix gun to a particle gun and adds helix surface intersections for planes and cylinders (see discussion on PR). The intersections are solved by the Newton method.
The old intersectors are renamed to ray intersectors for clarity and adapted to accept a ray as parameter. The ray and helix trajectories are added to the intersection folder.

Part of the PR is also a track state generator that reduces code duplication in the unittests. I also added const in a lot of places, too.