Add line mask and its intersector (#270)
This PR adds a mask for line (wire detector) and its corresponding intersector.

The line mask is defined with its half length and radial scope length. The intersection point is inside if its radial position is less than the radial scope length.
The line mask takes an additional template parameter of bool kSquareScope. The false case is for the straw tube where the detection scope is radial. The true case is for the drift chamber where the detection scope is square (I hope there is no rectangular case in the world...).
The intersection is obtained by calculating the line perpendicular to the track direction and line direction. The p2 of line intersection object contains a distance closest approach and position on the line.