Update CI container to version 26
This commit updates the version of the Acts container used by the CI to
version 26, from version 20. This gives us a newer version of CMake for
pull request #275. It also updates the git version to one with some more
security features which broke a small part of the benchmarking suite,
which was updated to accomodate for the changes.