Made the code set up the detray::Thrust target under all circumstances.
If Thrust is provided by the project that is building Detray as
a part of itself, we still need to create the detray::Thrust target.
(At that point the thrust_create_target(...) function needs to be
provided by the "parent project".)

This is in preparation of traccc making use of Thrust as well.