feat: add grid factory #346
Adds a dedicated grid factory that can be used to construct grids of different shapes. It is used in a grid builder class, that is later meant to decorate a volume builder. It builds an empty grid and then calls a fill function with a 'filler' functor to fill the grid.

In addition, there is some refactoring to the axis bounds in order to match ACTS definitions of axis shapes.