ref: move geometry and axis enums to definitions (#357)
Makes the type enums of axes, surfaces and volumes accessible from headers in the definition folder. Also add an enum for the volume shape, which is used to determine how to interpret the values in the volume bounds array.
It uses the detray internal enums for e.g. the axis labels directly in the json/io payloads and adds the id for the surface type, which needs to be supplied from the outside.