chore: clean out accessor header
Remove vtuple definition and old mask_store accessor from header, as they are not used anymore. The header is split into an algorithm header, a tuple_helper header and the fist_type trait and type_pos have been moved to type traits. Parts of the container definitions was moved to container_views.