Adds the material reader and connects it to the geometry reader by providing the read_detector function, which corresponds to the write_detector function. Since the detector builder works volume-by-volume, and the material building functionality is decorated onto the individual volume builders, the material io payloads had to be adapted to be read/written by volume, too.
In order to know which reader to register for a given file, the file header has
to be inspected. The header now contains a common part, that is e.g. used to
check if the file version is compatible with the running detray version, as well
as a subheader that is specific to every file. Currently the subheaders contain
some statistics about the file contant that can be used in the future to validate
the data or help prepare the reader infrastructure (e.g. reserving memory)