Add arbitrary bfield types (defined in a dedicated header) to the detector with comprehensive testing both for host and device. Since the covfie field types for host and device are a bit different, the detector and the detector_view had to be templated on the field/backend type, so that in the detector view and buffer classes the correct covfie constructors are called, which move the data to device. Because this needs a special interface for now, the detector has its own versions of the get_data and get_buffer functions. To make sure the detector with all the data it contains now is not copied, the copy constructor and copy-assignment operator were deleted and some downsteam adjustments for this were made in the io and simulation packages.
In order to manage the multiplicity of test cases (const bfield/inhomogeneous bfield) the stepper and
propagation unittests were adapted and new ones for the inhomogeneous bfield were added
(note: these can't compare against a helix anymore, since the tracks are no longer helical!).
The device propagation unittests involved a huge amount of duplicate code between the CUDA and
SYCL implementations which was moved into a common propagator test header (mostly the CPU code and
the comparison of the results between host and device). The sycl code that could not be unified
contains the sycl queue wrapper.

The toy detector has a new config struct now, which makes it easier to build different bfield types.
In many cases, the default values of this config are used and new values only set explicitely where
they diverge from the default.

The file handle saw some fixes, too. Mainly for the counting of open files and the correct handling of
empty file names.

A generic bfield reader has been added with a concrete reader implementation for covfie, which can now be
used to read the bfield files into the detector along the detray json data files.

The covfie version was bumped to 0.8.0

In order to make the CI pass, the ODD bfield file was uploaded to the traccc data repository and is
downloaded with the same script in detray now as is used in traccc

Co-authored-by: Stephen Nicholas Swatman <stephen.nicholas.swatman@cern.ch>
Co-authored-by: Guilherme Almeida <g.almeida@cern.ch>