ref: move surface vector into sf finder store #367
Since the current surface objects are essentially surface handles anyway, we can put them directly into the acceleration structures to save one level of indirection. The surface vector of the detector is now moved into the brute_force_finder, which is put into its own collection (one brute force finder per volume, which will later only hold the portals once the grid is fully ready).
This makes fetching the surfaces by index more difficult, but fetching them during navigation will be faster. A get_surface functor is needed, in order to retrieve the surface from the correct accelerator data structure instance. The multi_store functor interface has been refactored a bit, so that the return type does not need to be set explicitly anymore and void functors are possible as well.
The links that are held by the detector_volume are adapted as well: There is a link for every accelerator type onto which the surface types are matched (i.e. portal, sensitive ...), which also contains the index of the accelerator data structure instance in the collection that resides in the multi_store. All these links are iterated when filling the navigator cache.
Since a few problems came up, the Rule of Three was implemented for the detray::ranges and some of the tuple helper traits were simplified a bit.

Note: There is still a problem with the grid const correctness, which is why it was necessary to pull the grid types out of the toy geometry for now