perf: Remove Thrust calls inside kernels
@niermann999 and myself were profiling the propagation kernels in detray
and discovered significant overhead originating from calls to Thrust
functions in device code. In particular, a call to `thrust::sort` was
performing a sorting algorithm with non-constant spatial overhead,
meaning it had to allocate memory on the device side, which is quite
slow. This commit changes the sorting to use a hand-rolled
implementation of selection sort. Note that it also changes the use of
insertion sort to selection sort elsewhere, as insertion sort has an
average of O(n^2) swaps (which access global memory) compared to the
average of O(n) swaps that selection sort requires.

Co-authored-by: Joana Niermann <joana.niermann@cern.ch>