Refer to `nlohmann_json` target by namespace
This was a real pain in the ass to find and fix for
acts-project/traccc#364, but as it turns out `nlohmann_json` is not
actually a valid target, and the library has been exporting its targets
in a dedicated namespace since version 3.2. CMake expectedly handles
this poorly and translates it to a `-lnlohmann_json` flag if the target
is not available, leading to compilation errors. This commit applies the
namespace to resolve this issue.