Refactored detray::io into a single library.
For the moment removed the DETRAY_IO_CSV and DETRAY_IO_JSON
cache variables completely. They were not used correctly in
every place anyway. So for now the project would always need
dfelibs and nlohmann_json to be available.

In the move made sure that every I/O header would actually be
tested for working by itelf. Which was mistakenly not the
case before.