Fix `find_package` syntax for actsvg
When building detray with actsvg and when requesting a system
installation of actsvg to be used, the configuration of the project
currently fails. The reason for this is that the `find_package` command
is invoked with invalid syntax. More specifically, the current
invocation looks like the following:

```cmake
find_package( actsvg::core actsvg::meta REQUIRED )
```

However, requestion components from packages in this way is incorrect.
The correct invocation is the following:

```cmake
find_package( actsvg REQUIRED COMPONENTS core meta )
```

This commit updates the call to `find_package` such that CMake does not
reject it on sight.