ref: Reduce size of navigator memory alloc (#441)
Reduce the capacity of the navigation cache using the maximum number of candidates that a given volume may return. This is inferred from the neighborhood lookup of the acceleration data structure, but since they are omitted in the navigation right now, the number is fixed to 20 instead. For this reason, the telescope geometry will not allow more than 19 test surfaces for now.