This PR adds a json based writer for simple material (i.e. slabs and rods). (#438)
The json_writer now comprises the writers for both the geometry and simple material,
as well as future writers like the grid and material maps writers. An implementation of file
headers has been added, which contains general information like e.g. the detray version and
some limited statistics of the contained data. The file headers will not be mandatory during
reading as they cannot be fully generated by acts.

The writer interface code can potentially be generalized to an io interface for the readers to
avoid code duplication (including variadic parameter passing to the add function), but I will
leave that for the first reader PR. Also still open, tbh, is the question how best to expose
the file_handles in the user interface.