feat: json geometry reader #466
Adds a geometry reader from json files. The detector is built from the io payload using a vanilla volume builder that directly builds the volumes in the detector. This is a simplification for now compared to the full reader, where material and grids need to be added, too, before the detector can be built.

The geometry builder loops over the volume io payloads that are delivered by the json stream and initializes every detector volume according to its shape and boundaries. This automatically sets the correct volume index, since the json stream is ordered. It then proceeds to loop over the surface payloads contained in the volume payloads and creates a surface factory per shape (e.g. ractangle, cylinder...) and surface type (portal, sensitive, passive). This is done via sfinae, so that the usage of masks types or enum elements that are unknown to a given detector does not cause an error. The surface factories are kept in a map and can be retrieved for every new surface payload using a key that contains the respective shape and surface type ids. In this way, every surface payload will be added to the correct surface factory. Once the factories are filled, they are passed to the volume builder, where their data is transcribed into the builders data stores. These data stores are then appended to the detector in the last step and all links that the volume builder created are updated accordingly.

Note: I removed the polymorphism on the surface_data type, which is meant to securely strap the data for a single surface instance together, so that is is always strongly associated with its data components throughout the building process. I might need to go back on that when I add the material builder decorator, but for now this way was easier. Some of that code remains as comments until the full thing is implemented.

The toy detector unittest was moved into its own header, so that I could reuse it to do a first validation of the result from the reader. The reader test does the full roundtrip: create a toy detector -> write it to file -> read it back in -> run the toy detector unittest on it. For the test to pass, the beampipe in the toy detector had to be constructed after the portals, to match the behavior of the volume builder. Additionally, the material link checking is temporarily disabled, until the material can be handled by the io.