ref: add buffers to existing detray stores (#488)
This PR enables to copy the detector to device with minimal code changes. This is probably not the targeted design, but at least it will be much easier to set up as a first solution. The code is also still pretty raw and the bfield and copy synchronization will need to be addressed in a followup.

Most of what lives in container_buffers.hpp are type traits, which are not strictly needed, but will make the thing a bit safer to use. The implementation itself works the same way as the recursive view construction that we already have. It treats the buffers as an intermediate memory representation that mimics the hierarchy of member variables in our classes, but contains no functionality on its own. The view on these buffers is likewise built up recursively and then passed to the interface types (the same classes, but using vecmem::device_vector) which will get constructed around them the same way they are now, by handing the view types through the chain of constructors. While recursions always need a bit of mental yoga to understand, it will safe us from a lot of boilerplate code of defining host, view and buffer classes for all of our memory handling types. Also, we can keep using all of the managed memory setups in parallel, so there will be no downstream changes.