Similar to the volumes, this PR adds an interface around the detector surfaces that allows us to pool the functionality of a surface. This also provides a layer of abstraction around the surfaces descriptors, so that users and downstream projects don't have to handle e.g. transform or mask links directly.
The old surface class has been renamed to surface_descriptor and two new surface classes
are added that implement the interface. One only contains geometric functionality
(e.g. local-to-global transformation) and the other extends this by propagation functionality
(e.g. bound-to-free Jacobian). The reason they are split is, that this allows to compile detray
with only geometry and navigation functionality (which is desired for ray tracing/Celeritas).
The corresponding functions on the detector have been removed to keep responsibilities cleanly divided.
The rest of the changes are the adoption of the new interface.