Adds transforms to the volumes, so that rotations and translations of the volume can be properly included when switching to grid local coordinates for the axes lookup during local navigation. The volume transforms are the first to be added to the transform store respectively and are non contextual (don't need alignment) like the portal transforms that follow immediately after. The volume transforms are also added by a few new methods (add_volume_placement) to the volume builders.
Since the transform indices change with the extra transforms in the transform store,
the unittests have been adapted. Furthermore, the index updates in the grid filling
for the toy detector was simplified/corrected (e.g. global_to_local in the bin filler
can now use proper transform) and the volume descriptor was cleaned up a bit.