This adds the grids to the navigation, but does not perform local navigation, yet. (#456)
Instead, all grid bins are tested for now by flattening the grid using a join iterator.
Join will concatenate the iterations over every grid bin into a single iterator.

In order to test the setup the cylinder and disc grids in the toy detector are reactivated and the unittests
adapted accordingly.

Unfortunately, keeping the surface descriptors in the accelerator data structures uniquely did not work
(since they might be placed multiple times, leading to duplicates that can later not be told apart and
because accessing them there becomes quite complicated). Therefore, I added the surface vector back in the detector.