ref: split metadata and add full metadata as default to detector (#486)
For traccc a full detector type should be used that can hold any type of shape or grid that is available in detray. I split off the full_metadata that is meant to do this from the utils directory and added it to core as a default argument for the detector, so that the comprehensive detector type can be buil as detray::detector<> .
I added the debug shape types that are only used in detray, so we might want to remove them again from the core type if we continue to use the dedicated telescope type in detray. Currently, I also did not add all grid types that we may get as input from ACTS (all possible binning and boundary type combinations). Once they come up, we need to add them at that point.
This also adds a test of reading in the toy detector from json into the detector type with the full metadata