ref: some fixes for ITk (#478)
Adds a few fixes that were needed in order to read the preliminary ITk json file. Mostly, the enumeration of the annulus shape had to be adapted to be exactly like it is defined in ACTS and an ITk detector metadata file was added. There is no dedicated ITk IO executable though.