Add concurrency group for CI jobs
As it stands, an arbitrary number of CI jobs can be running for a given
pull request, which means that sometimes the CI will be running for a
commit that has been superceded by several newer commits. Adding a
concurrency group to the CI jobs allows GitHub to automatically cancel
old jobs if a new commit is pushed that makes the run obsolete.