Add a number of utility functions that perform some simple consistency checks on a detector object and its volumes and surfaces:
  -  check for empty data stores
  -  check for any invalid links in volumes and surfaces
  -  check that volumes and surfaces have the correct volume index set
  -  check that mask boundaries are consistent
  -  check that sensitive/passive and portals surfaces have a navigation link either to their mother volume or not, respectively
  -  check that every surface in the brute force method/grids can be found via the detector lookup and vice versa

To avoid a segfault during the checking run, I added an empty (and in the IO a dummy) volume grid to the detector.
This will be handled properly in a future PR.

To check different detectors read in from json, I added a folder called checks, where
all of the consistency check will be called in the future. There is a test now that
inherits from the google test test via a new default test fixture, which call the new
consistency check.