feat: implement surface normal and incidence angle (#504)
This add the access to the surface normal (both for bound positions and local positions) to the surface interface, as well as the calculation of the incidence angle theta. This will allow to compute the material interactions from bound track parameters without calling an intersector to get cos(theta).

Also adds and extends unittest for both the surface and surface_descriptor classes.