feat: grid json writer (#510)
This PR add a generic grid writer and its json extension to the detector writer. The grid writer will go through the acceleration data structure store (surface_store) of the detector, examine at compile time if a collection contains grids and serialize each grid of the collection into an IO payload. The grid payload contains an axis payload (with enums for binning, bounds etc) and a bin payload with the bin content an a local bin index. The bin index has the same ordering as the axis payload, i.e. the first index is the index corresponding to the first axis and so on. The axes are labeled additionally from 0 to the grid dimension. Currently, the bin content can only be indices, which correspond to the global surface indices.

The grid IO payloads have been extended to hold grid collection, bins and the grid file header. Unittests for this were added, too.

Edit: The acceleration data structure type is now also correctly set for the geometry writer, instead of just being cast to the correct type while using a detector specific value

Edit: The writers now use the type_registry to match the IO enums to the detray types