feat: detector builder (#516)
Adds a detector builder that holds the volume builders for a given detector in a specific data structure (currently flat vector). When all of the data for the single volumes is accumulated, either generated or read in, then the full detector is assembled. Here, data de-duplication and sorting will be implemented in the future. If additional detector components, like e.g. material, need to be added, the volume builders can be decorated with specialized volume builders that will add the required component during the building step.