Add material scan and corresponding plotting tools. The material scanner is added to the validations suite and can shoot rays in uniform eta steps through a detector and record the total encountered material for each ray. The material is recorded either per radiation or interaction length and as total thickness of the material vs. ray path length through the material.
In a second tool (implemented in python), the resulting material budget
can be plotted automatically, either as 2D eta-phi map or as a histogram
against eta. The plotting functionality has been abstracted into a
plot_factory to be reusable in other workflows.

The chain can be run with the following commands:

> ./bin/detray_material_validation --geometry_file toy_detector_geometry.json --material_file toy_detector_homogeneous_material.json --phi_steps 500 --eta_steps 500

and the plotting:

> python3 ../tests/validation/python/run_material_validation.py --input material_scan_toy_detector.csv --output_format pdf

Help messages are also available