feat: implement plotting of ray scan positions in python (#567)
Add functionality to output the positions of the ray surface intersections, as well as python tools to plot them in an xy - and an rz view. The data can both be plotted for the ray and the helix scan, which needs to be turned on by the --write_scan_data option.