feat: material maps writer (#587)
This implements the material maps writing. It splits the common writer functionality 
for the grids in a common helper class that is inherited by the surface_grid_writer 
and the new material_map_writer. The material_map_writer uses the grid_payload
with a material_slab_payload as value type.