feat: grid search (#603)
Implement the neighborhood lookup on the grids using in-place iteration (no memory allocations for a result vector):

    lookup of the index ranges on the axes that correspond to the search position plus a search window around it
    generate an index sequence for every local bin axis range
    in order to index every bin in the search area, calculate the cartesian product of all of the index sequences
    map the cartesian product to global bin indices
    fetch the bin, which is itself a range
    join all of the bin ranges into a single iteration

Except for the first and the last step, everything else is handled by the new bin_view class, which can also be used in other contexts than local navigation to iterate through the bins in a neighborhood of a grid.

For the phi axes, which have periodic boundary conditions for their bin indices, the wrapping has been separated from the mapping of the search window on the axis range, so that the generation of the index sequence is well defined. The resulting bin index sequence is mapped onto the periodic boundary conditions in a dedicated step during the bin view iteration.