Some fixes for ITk (#604)
Increase the grid bin capacity to fit all candidates and fix vertex generation sign bug. Additionally, it restructures the grid building a little, so that the contract with the volume builder is that a boolean needs to be set when a builder for an acceleration structure is present. This allows for better reasoning on how and when to fill the brute force navigation method with surfaces