feat: Implement surface lookup from source link #576
This abstracts the surface lookup from the detector class, so that it can be extended later on without breaking anything. The surface lookup is renamed to just surfaces again. It also strips the source link from the surface descriptor, which will now not be carried around throughout all of the navigation and grid. The source link is still available via the surface lookup.

In order to make easier lookups for the source links, the portal, sensisitve and passive index ranges are added back to the volume descriptor, which also allows to get access to the surface range of a volume via the detector_volume class. Wherever I could find it, I already refactored the surface access through the volume accordingly.