ref: Move axes definitions out of shapes and renaming (#625)
This moves the axes definitions out of the shapes, hence decoupling the grids from the geometry, and adds a few more helper traits that make grid type definitions a bit simpler.

Renames axis (the current axis namespace of grid2) to axis2 and n_axis to axis. Harmonizes the capitalization of the dim parameter over the classes that define a local dimension.

Adds a few more convenience grid creation overloads to the grid factory, which are used in the grid builder.

Removes an unnecessary typedef on the scalar_t from the stepper.