Unit test clean up (#664)
Sort the test files into subdirectories and add the tests in core and utils to the unit test directory. Also merge the tests_common and detray::test libraries, as well as the svgtools executables. Some tests were split or reordered for easier readability, plus a bit of general clean up.

It seems, that grid2 was never tested against any algebra-plugins. In CUDA, the complete populator fails, since the operator= is not constexpr for std::array in c++17. The compilation with Eigen failed and was hence skipped.