ref: remove abstract interface from backend io  (#668)
Remove the io_interface from the backend readers and writers and simplify the templating. Renames the methods of the backend readers and writers to convert and splits some headers up, so that reading and writing are now separated completely. The type definitions of the full reader/writer implementation is now done in the frontend and only if the detector type under construction contains the respective component. This might make the compilation for e.g. telescope detectors a bit easier.