Add volume material types back and avoid using vacuum material (#687)
In order to finalize the material file format, reactivate the volume material. All volume material is now properly put into the material store and referenced through the volume descriptor. In order to access the different types of materials, a material_accessor was added. Also adds full material consistency checks in the detector IO.