ref: decouple propagation and geometry (#674)
This is factored out from #555, to make the vetcorized intersectors compile and decouple the dependency between the jacobian transport implementation and the base geometry building blocks.

Since the geometry describes abstract shapes that are used for more than the detector surfaces, compiling the geometry on top of the track parametrization and the jacobian implementation is very cumbersome. This PR switches that dependency around, so that instead the track parametrization transformation and the generation of the jacobians is compiled using the geometry, which allows for way more freedom in the geometry description. E.g. neither bounding volumes nor grids, nor the navigation depend on anything anymore that is only used in the propagation.
In the propagation code, the main change is that the bound_to_free_jacobian etc are called from the jacobian_engine instead of the coordinate frame type. The track parametrization transformations are free functions now that only need the detail namspace in front. The way the calculations are done is not changed. The definitions of the free_vector etc has been centralized, so that any changes can be applied in a single place.

Nothing is changed conceptually, only where and when it is getting compiled